move serialization to utils
